+++ runtime/core/test/function_ref_test.cpp
