+++ test/utils/OSSTestConfig.json
@@ -45,7 +45,6 @@
     {
         "directory": "extension/pytree/test",
         "sources": [
-            "function_ref_test.cpp",
             "test_pytree.cpp"
         ]
     },
@@ -96,14 +95,15 @@
     {
         "directory": "runtime/core/test",
         "sources": [
-            "span_test.cpp",
+            "array_ref_test.cpp",
             "error_handling_test.cpp",
+            "evalue_test.cpp",
             "event_tracer_test.cpp",
             "freeable_buffer_test.cpp",
-            "array_ref_test.cpp",
-            "memory_allocator_test.cpp",
+            "function_ref_test.cpp",
             "hierarchical_allocator_test.cpp",
-            "evalue_test.cpp"
+            "memory_allocator_test.cpp",
+            "span_test.cpp"
         ]
     },
     {
